+++ energyflow/utils/data_utils.py
@@ -1,8 +1,10 @@
-"""## Data Tools
+"""## Data Utils
 
-Functions for dealing with datasets. These are not importable from
-the top level `energyflow` module, but must instead be imported 
-from `energyflow.utils`.
+Functions for handling with datasets, including facilitating [train/val/test 
+splits](#data_split), [converting](#convert_dtype) the numpy dtype of a
+(possibly ragged) array, [padding events](#pad_events) with different numbers
+of particles, and [mapping](#remap_pids) PDG ID values to small floating point
+values.
 """
 
 #  _____       _______               _    _ _______ _____ _       _____
@@ -24,11 +26,15 @@ import sys
 import numpy as np
 from six.moves.urllib.error import HTTPError, URLError
 
+from energyflow.utils.generic_utils import kwargs_check
 from energyflow.utils.random_utils import random
 
 __all__ = [
-    'get_examples', 
-    'data_split', 
+    'convert_dtype',
+    'data_split',
+    'determine_cache_dir',
+    'get_examples',
+    'pad_events',
     'to_categorical',
     'remap_pids'
 ]
@@ -48,18 +54,158 @@ ALL_EXAMPLES = [
     'animation_example.py'
 ]
 
-def get_examples(cache_dir='~/.energyflow', which='all', overwrite=False):
+def convert_dtype(X, dtype):
+    """Converts the numpy dtype of the given array to the provided value. This
+    function can handle a ragged array, that is, an object array where the
+    elements are numpy arrays of a possibly different type, in which case the
+    function will be recursively applied.
+
+    **Arguments**
+
+    **Returns**
+
+    - _numpy.ndarray_
+
+    """
+
+    # if dtype is None, do nothing
+    if dtype is None:
+        return X
+
+    # check for proper argument type
+    if not isinstance(X, np.ndarray):
+        raise TypeError('argument must be a numpy ndarray')
+
+    # object arrays are special
+    if X.dtype == 'O':
+        return np.asarray([convert_dtype(x, dtype) for x in X], dtype='O')
+    else:
+        return X.astype(dtype, copy=False)
+
+# data_split(*args, train=-1, val=0.0, test=0.1, shuffle=True, include_empty=False)
+def data_split(*args, **kwargs):
+    """A function to split a dataset into train, validation, and test datasets.
+
+    **Arguments**
+
+    - ***args** : arbitrary _numpy.ndarray_ datasets
+        - An arbitrary number of datasets, each required to have the same number
+        of elements, as numpy arrays.
+    - **train** : {_int_, _float_}
+        - If a float, the fraction of elements to include in the training set.
+        If an integer, the number of elements to include in the training set.
+        The value `-1` is special and means include the remaining part of the
+        dataset in the training dataset after the test and (optionally) val
+        parts have been removed.
+    - **val** : {_int_, _float_}
+        - If a float, the fraction of elements to include in the validation set.
+        If an integer, the number of elements to include in the validation set.
+        The value `0` is special and means do not form a validation set.
+    - **test** : {_int_, _float_}
+        - If a float, the fraction of elements to include in the test set. If an
+        integer, the number of elements to include in the test set. The value `0`
+        is special and means do not form a validation set.
+    - **shuffle** : _bool_
+        - A flag to control whether the dataset is shuffled prior to being split
+        into parts.
+    - **include_empty** : _bool_
+        - Whether or not to return empty arrays for datasets that would have
+        zero elements in them. This can be useful for setting e.g. `val` or
+        `test` to 0 without having to change the unpacking of the result.
+
+    **Returns**
+
+    - _list_
+        - A list of the split datasets in train, val, test order. If datasets
+        `X`, `Y`, and `Z` were given as `args` (and assuming a non-zero `val`
+        and `test`), then [`X_train`, `X_val`, `X_test`, `Y_train`, `Y_val`,
+        `Y_test`, `Z_train`, `Z_val`, `Z_test`] will be returned. If, for
+        instance, `val` is zero and `include_empty` is `False` then [`X_train`,
+        `X_test`, `Y_train`, `Y_test`, `Z_train`, `Z_test`] will be returned.
+    """
+
+    # handle valid kwargs
+    train, val, test = kwargs.pop('train', -1), kwargs.pop('val', 0.0), kwargs.pop('test', 0.1)
+    shuffle, include_empty = kwargs.pop('shuffle', True), kwargs.pop('include_empty', False)
+    kwargs_check('data_split', kwargs)
+
+    # validity checks
+    if len(args) == 0: 
+        raise ValueError('need to pass at least one positional argument to data_split')
+
+    # check for consistent length and types
+    n_samples = len(args[0])
+    for arg in args[1:]:
+        if len(arg) != n_samples:
+            raise ValueError('args to data_split have different lengths')
+        if not isinstance(arg, np.ndarray):
+            raise TypeError('args must be numpy.ndarray instances')
+
+    # determine numbers
+    num_val = int(n_samples*val) if val <= 1 else val
+    num_test = int(n_samples*test) if test <= 1 else test
+    num_train = (n_samples - num_val - num_test
+                 if train == -1
+                 else (int(n_samples*train) if train <= 1 else train))
+
+    # sensibility checks
+    if num_train < 0:
+        raise ValueError('calculated num_train is negative')
+    if num_train + num_val + num_test > n_samples:
+        raise ValueError('too few samples for requested data split')
+    
+    # calculate masks 
+    perm = random.permutation(n_samples) if shuffle else np.arange(n_samples)
+    masks = []
+    dset_func = lambda n, s: masks.append(perm[s]) if n > 0 or include_empty else None
+    dset_func(num_train, slice(0, num_train))
+    dset_func(num_val,   slice(num_train, num_train+num_val))
+    dset_func(num_test,  slice(num_train+num_val, num_train+num_val+num_test))
+
+    # return list of new datasets
+    return [arg[mask] for arg in args for mask in masks]
+
+def determine_cache_dir(cache_dir=None, cache_subdir='datasets'):
+    """Determines the path to the specified directory used for caching files. If
+    `cache_dir` is `None`, the default is to use `'~/.energyflow'` unless the
+    environment variable `ENERGYFLOW_CACHE_DIR` is set, in which case it is
+    used.
+
+    **Arguments**
+
+    - **cache_dir** : _str_ or `None`
+        - The path to the top-level cache directory. Defaults to the environment
+        variable `ENERGYFLOW_CACHE_DIR`, or `'~/.energyflow'` if that is unset.
+    - **cache_subdir** : _str_ or `None`
+        - Further path component to join to `cache_dir`. Ignored if `None`.
+
+    **Returns**
+
+    - _str_
+        - The path to the cache directory specified by the supplied arguments.
+    """
+
+    # translate cache_dir as None to default
+    if cache_dir is None:
+        cache_dir = os.environ.get('ENERGYFLOW_CACHE_DIR', '~/.energyflow')
+
+    # consider appending cache_subdir
+    if cache_subdir is not None:
+        cache_dir = os.path.join(cache_dir, cache_subdir)
+
+    return os.path.expanduser(cache_dir)
+
+def get_examples(cache_dir=None, which='all', overwrite=False):
     """Pulls examples from GitHub. To ensure availability of all examples
     update EnergyFlow to the latest version.
 
     **Arguments**
 
-    - **cache_dir** : _str_
-        - The directory where to store/look for the files. Note that 
-        `'examples'` is automatically appended to the end of this path. If
-        `None`, the default path of `~/.energyflow` is used, unless the
-        environment variable `ENERGYFLOW_CACHE_DIR` is set in which case that
-        is used instead.
+    - **cache_dir** : _str_ or `None`
+        - The directory where to store/look for the files. If `None`, the
+        [`determine_cache_dir`](../utils/#determine_cache_dir) function will be
+        used to get the default path. Note that in either case, `'datasets'` is
+        appended to the end of the path.
     - **which** : {_list_, `'all'`}
         - List of examples to download, or the string `'all'` in which 
         case all the available examples are downloaded.
@@ -79,7 +225,7 @@ def get_examples(cache_dir='~/.energyflow', which='all', overwrite=False):
         examples = all_examples.intersection(which)
 
     base_url = 'https://github.com/pkomiske/EnergyFlow/raw/master/examples/'
-    cache_dir = os.path.expanduser(_determine_cache_dir(cache_dir))
+    cache_dir = determine_cache_dir(cache_dir)
 
     # get each example
     files = []
@@ -100,77 +246,23 @@ def get_examples(cache_dir='~/.energyflow', which='all', overwrite=False):
         print(fname, 'exists at', path)
     print()
 
-# data_split(*args, train=-1, val=0.0, test=0.1, shuffle=True)
-def data_split(*args, **kwargs):
-    """A function to split a dataset into train, test, and optionally 
-    validation datasets.
+def pad_events(X, pad_val=0., max_len=None):
+    """"""
 
-    **Arguments**
+    if max_len is None:
+        max_len = max([len(x) for x in X])
 
-    - ***args** : arbitrary _numpy.ndarray_ datasets
-        - An arbitrary number of datasets, each required to have
-        the same number of elements, as numpy arrays.
-    - **train** : {_int_, _float_}
-        - If a float, the fraction of elements to include in the training
-        set. If an integer, the number of elements to include in the
-        training set. The value `-1` is special and means include the
-        remaining part of the dataset in the training dataset after
-        the test and (optionally) val parts have been removed
-    - **val** : {_int_, _float_}
-        - If a float, the fraction of elements to include in the validation
-        set. If an integer, the number of elements to include in the
-        validation set. The value `0` is special and means do not form
-        a validation set.
-    - **test** : {_int_, _float_}
-        - If a float, the fraction of elements to include in the test
-        set. If an integer, the number of elements to include in the
-        test set.
-    - **shuffle** : _bool_
-        - A flag to control whether the dataset is shuffled prior to
-        being split into parts.
+    output_shape = (len(X), max_len, X[0].shape[1])
+    if pad_val == 0.:
+        output = np.zeros(output_shape, dtype=X[0].dtype)
+    else:
+        output = np.full(output_shape, pad_val, dtype=X[0].dtype)
 
-    **Returns**
+    # set events in padded array
+    for i, x in enumerate(X):
+        output[i,:len(x)] = x
 
-    - _list_
-        - A list of the split datasets in train, [val], test order. If 
-        datasets `X`, `Y`, and `Z` were given as `args` (and assuming a
-        non-zero `val`), then [`X_train`, `X_val`, `X_test`, `Y_train`, 
-        `Y_val`, `Y_test`, `Z_train`, `Z_val`, `Z_test`] will be returned.
-    """
-
-    # handle valid kwargs
-    train, val, test = kwargs.pop('train', -1), kwargs.pop('val', 0.0), kwargs.pop('test', 0.1)
-    shuffle = kwargs.pop('shuffle', True)
-    if len(kwargs):
-        raise TypeError('following kwargs are invalid: {}'.format(kwargs))
-
-    # validity checks
-    if len(args) == 0: 
-        raise RuntimeError('Need to pass at least one argument to data_split')
-
-    # check for consistent length
-    n_samples = len(args[0])
-    for arg in args[1:]: 
-        assert len(arg) == n_samples, 'args to data_split have different length'
-
-    # determine numbers
-    num_val = int(n_samples*val) if val<=1 else val
-    num_test = int(n_samples*test) if test <=1 else test
-    num_train = n_samples - num_val - num_test if train==-1 else (int(n_samples*train) if train<=1 else train)
-    assert num_train >= 0, 'bad parameters: negative num_train'
-    assert num_train + num_val + num_test <= n_samples, 'too few samples for requested data split'
-    
-    # calculate masks 
-    perm = random.permutation(n_samples) if shuffle else np.arange(n_samples)
-    train_mask = perm[:num_train]
-    val_mask = perm[-num_val:]
-    test_mask = perm[num_train:num_train+num_test]
-
-    # apply masks
-    masks = [train_mask, val_mask, test_mask] if num_val > 0 else [train_mask, test_mask]
-
-    # return list of new datasets
-    return [arg[mask] for arg in args for mask in masks]
+    return output
 
 def to_categorical(labels, num_classes=None):
     """One-hot encodes class labels.
@@ -238,7 +330,6 @@ def remap_pids(events, pid_i=3, error_on_unknown=True):
             events[:,:,pid_i] = np.asarray([PID2FLOAT_MAP.get(pid, 0.)
                                             for pid in pids]).reshape(events.shape[:2])
 
-
     # single event
     elif events.ndim == 2:
         pids = events[:,pid_i].astype(int)
@@ -273,12 +364,7 @@ def _pad_events_axis1(events, axis1_shape):
 
     return events
 
-def _determine_cache_dir(cache_dir):
-    if cache_dir is None:
-        cache_dir = os.environ.get('ENERGYFLOW_CACHE_DIR', '~/.energyflow')
-    return cache_dir
-
-# the following code is closely based on analogous parts of Keras
+# the following code is based on analogous parts of Keras
 if sys.version_info[0] == 2:
     from contextlib import closing
     from six.moves.urllib.request import urlopen
@@ -366,23 +452,29 @@ def _validate_file(fpath, file_hash, algorithm='auto', chunk_size=131071):
 
     return str(_hash_file(fpath, hasher, chunk_size)) == str(file_hash)
 
-def _get_filepath(filename, url, cache_dir, cache_subdir='datasets', file_hash=None):
+# PTK: modified this function given our `determine_cache_dir` function above
+def _get_filepath(filename, url, cache_dir=None, cache_subdir='datasets', file_hash=None):
     """Pulls file from the internet."""
 
-    # handle '~' in path
-    datadir_base = os.path.expanduser(cache_dir)
+    # get cache_dir
+    datadir = determine_cache_dir(cache_dir, cache_subdir)
 
-    # ensure that directory exists
-    datadir = os.path.join(datadir_base, cache_subdir)
-    if not os.path.exists(datadir):
-        os.makedirs(datadir)
-
-    # handle case where cache is not writeable
-    if not os.access(datadir_base, os.W_OK):
-        datadir = os.path.join('/tmp', '.energyflow', cache_subdir)
+    # ensure that directory exists and is writable
+    try:
         if not os.path.exists(datadir):
             os.makedirs(datadir)
 
+        if not os.access(datadir, os.W_OK):
+            raise PermissionError()
+
+    # need to change to backup location
+    except PermissionError:
+        datadir = os.path.join('/tmp', '.energyflow',
+                               cache_subdir if cache_subdir is not None else '')
+        if not os.path.exists(datadir):
+            os.makedirs(datadir)
+
+    # get filepath
     fpath = os.path.join(datadir, filename)
 
     # determine if file needs to be downloaded