+++ energyflow/utils/random_utils.py
@@ -1,4 +1,4 @@
-"""## Random Events
+"""## Random Utils
 
 Functions to generate random sets of four-vectors. Includes an implementation
 of the [RAMBO](https://doi.org/10.1016/0010-4655(86)90119-0) algorithm for
@@ -32,6 +32,14 @@ if hasattr(np.random, 'default_rng'):
     random = np.random.default_rng()
 else:
     random = np.random
+
+# random
+def random4doc():
+    """In NumPy versions >= 1.16, this object is obtained from
+    `numpy.random.default_rng()`. Otherwise, it's equivalent to `numpy.random`.
+    """
+
+    pass
 
 def gen_random_events(nevents, nparticles, dim=4, mass=0.):
     r"""Generate random events with a given number of particles in a given